Add performance tests for vdom.
These tests show that vdom techniques work kind of
as expected--nothing is super expensive.

The tests use 300 persons in the PM list, which is
way more than typical users, but everything only
takes a few milliseconds.

Here are some results (not in the same order that the
tests execute):

Here is what happens in the update case.  Basically,
it only takes 2ms to detect that nothing has actually
changed in our vdom, and we will either do nothing or
just re-render one list item:

    equality checks
    elapsed (ms) 201
    per (ms) 2.01

    vdom.update
    elapsed (ms) 208
    per (ms) 2.08

    vdom.update w/one change
    elapsed (ms) 228
    per (ms) 2.28

Also building the vdom is pretty cheap (this is
just building the AST from pm_list data that has
already been calculated:

    build vdom
    elapsed (ms) 130
    per (ms) 1.3

For contrast, we actually do more work to calculate
the data that we would normally stuff into a template.
Some of the expensive things there are calculating
hash urls and figuring out presence:

    get convos (before stubbing)
    elapsed (ms) 382
    per (ms) 3.82

    get convos (w/stubbed unread)
    elapsed (ms) 340
    per (ms) 3.4

    get convos (w/more stubbing)
    elapsed (ms) 112
    per (ms) 1.12

None of the above are super expensive--it's to be
expected when you have about 300 items in a list
that you might do a few milliseconds of work.

All of the rest of the numbers are more full-stack
stuff (and mostly are expensive due to things
we already pay every time we update PMs):

    just rendering
    elapsed (ms) 588
    per (ms) 5.88

    close/expand
    elapsed (ms) 1001
    per (ms) 10.01

    update with no changes
    elapsed (ms) 857
    per (ms) 8.57

    updates with unread change
    elapsed (ms) 818
    per (ms) 8.18