internal: Add internal include_user_streams logic.
We add a bool option to gather_subscriptions_helper called
include_user_streams and a by_user flag to
bulk_get_subscriber_user_ids.

This internal change sets us up to decrease the size of
subscribers in our page_params payload by formatting
subscribers as a dictionary of users to streams.  (Since
most users have fairly similar streams, normal compression
algorithms work better with the underlying matrix of
streams/users transposed to users/streams.)

We should wait to merge this until subsequent commits that
open up the API to client, but we won't want to actually
squash this, since we may decide to change the outward API
in the near future. This internal code is likely to
remain valid even if we decide to do some fancy lazy-logic
or something.

One of the things that is slightly irritating about this
change is that I don't rename the `include_subscribers`
flag, which now has the more precise semantics of
include-subscribers-in-sub-records.  The problem is that
we have to maintain that flag in our outward API.  For
now I just assert that `include_subscribers` and
`include_user_streams` are mutually exclusive options.
Down the road we might change the code so that only
the boundary code (i.e. the view) uses
`include_subscribers` and then we pass down something
more clear to the internal helpers.