performance: Use user_streams for subscribers.
We now send subscribers to the webapp in
the `user_streams` data structure that
maps users to streams, not vice versa.

Representing subscribers as user_id -> stream_ids
compresses about twice as well as our old
method of sub.subscribers -> user_ids, based
on real world data for czo.

See https://chat.zulip.org/#narrow/stream/
    3-backend/topic/subscribers.20perf/near/804894

Note that this commit does not yet change
the external API for registering for events;
instead, it's an undocumented feature for
the webapp.  We will probably want to open up
the API to at least zulip-terminal (ZT) and
then document the feature.